New kas_core::theme module
