+++ crates/kas-core/src/core/impls.rs
@@ -6,10 +6,10 @@
 //! Trait impls
 
 use super::*;
-use crate::draw::{DrawHandle, SizeHandle};
 use crate::event::{self, Event, Manager, ManagerState, Response};
 use crate::geom::{Coord, Rect};
 use crate::layout::{AlignHints, AxisInfo, SizeRules};
+use crate::theme::{DrawHandle, SizeHandle};
 use crate::{CoreData, WidgetId};
 use std::any::Any;
