+++ crates/kas-core/src/theme/draw.rs
@@ -8,262 +8,11 @@
 use std::convert::AsRef;
 use std::ops::{Bound, Deref, DerefMut, Range, RangeBounds};
 
+use super::{InputState, SizeHandle, TextClass};
 use crate::dir::Direction;
 use crate::draw::{color::Rgb, Draw, ImageId, PassType};
-use crate::geom::{Coord, Offset, Rect, Size};
-use crate::layout::{AxisInfo, FrameRules, Margins, SizeRules};
+use crate::geom::{Coord, Offset, Rect};
 use crate::text::{AccelString, Text, TextApi, TextDisplay};
-
-// for doc use
-#[allow(unused)]
-use crate::text::TextApiExt;
-
-bitflags! {
-    /// Input and highlighting state of a widget
-    ///
-    /// This struct is used to adjust the appearance of [`DrawHandle`]'s primitives.
-    #[derive(Default)]
-    pub struct InputState: u8 {
-        /// Disabled widgets are not responsive to input and usually drawn in grey.
-        ///
-        /// All other states should be ignored when disabled.
-        const DISABLED = 1 << 0;
-        /// Some widgets, such as `EditBox`, use a red background on error
-        const ERROR = 1 << 1;
-        /// "Hover" is true if the mouse is over this element
-        const HOVER = 1 << 2;
-        /// Elements such as buttons, handles and menu entries may be depressed
-        /// (visually pushed) by a click or touch event or an accelerator key.
-        /// This is often visualised by a darker colour and/or by offsetting
-        /// graphics. The `hover` state should be ignored when depressed.
-        const DEPRESS = 1 << 3;
-        /// Keyboard navigation of UIs moves a "focus" from widget to widget.
-        const NAV_FOCUS = 1 << 4;
-        /// "Character focus" implies this widget is ready to receive text input
-        /// (e.g. typing into an input field).
-        const CHAR_FOCUS = 1 << 5;
-        /// "Selection focus" allows things such as text to be selected. Selection
-        /// focus implies that the widget also has character focus.
-        const SEL_FOCUS = 1 << 6;
-    }
-}
-
-impl InputState {
-    /// Extract `DISABLED` bit
-    #[inline]
-    pub fn disabled(self) -> bool {
-        self.contains(InputState::DISABLED)
-    }
-
-    /// Extract `ERROR` bit
-    #[inline]
-    pub fn error(self) -> bool {
-        self.contains(InputState::ERROR)
-    }
-
-    /// Extract `HOVER` bit
-    #[inline]
-    pub fn hover(self) -> bool {
-        self.contains(InputState::HOVER)
-    }
-
-    /// Extract `DEPRESS` bit
-    #[inline]
-    pub fn depress(self) -> bool {
-        self.contains(InputState::DEPRESS)
-    }
-
-    /// Extract `NAV_FOCUS` bit
-    #[inline]
-    pub fn nav_focus(self) -> bool {
-        self.contains(InputState::NAV_FOCUS)
-    }
-
-    /// Extract `CHAR_FOCUS` bit
-    #[inline]
-    pub fn char_focus(self) -> bool {
-        self.contains(InputState::CHAR_FOCUS)
-    }
-
-    /// Extract `SEL_FOCUS` bit
-    #[inline]
-    pub fn sel_focus(self) -> bool {
-        self.contains(InputState::SEL_FOCUS)
-    }
-}
-
-/// Class of text drawn
-///
-/// Themes choose font, font size, colour, and alignment based on this.
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Ord, PartialOrd, Hash)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-pub enum TextClass {
-    /// Label text is drawn over the background colour
-    Label,
-    /// Scrollable label (same as label except that min height is limited)
-    LabelScroll,
-    /// Button text is drawn over a button
-    Button,
-    /// Class of text drawn in a single-line edit box
-    Edit,
-    /// Class of text drawn in a multi-line edit box
-    EditMulti,
-    /// Menu label (single line, does not stretch)
-    MenuLabel,
-}
-
-impl TextClass {
-    /// True if text should be automatically line-wrapped
-    pub fn line_wrap(self) -> bool {
-        self == TextClass::Label || self == TextClass::EditMulti
-    }
-}
-
-/// Default class: Label
-impl Default for TextClass {
-    fn default() -> Self {
-        TextClass::Label
-    }
-}
-
-/// A handle to the active theme, used for sizing
-///
-/// The shell provides widgets a `&dyn SizeHandle` in [`crate::Layout::size_rules`].
-/// It may also be accessed through [`crate::event::Manager::size_handle`] and
-/// [`DrawHandle::size_handle`].
-///
-/// All methods get or calculate the size of some feature.
-///
-/// See also [`DrawHandle`].
-pub trait SizeHandle {
-    /// Get the scale (DPI) factor
-    ///
-    /// "Traditional" PC screens have a scale factor of 1; high-DPI screens
-    /// may have a factor of 2 or higher; this may be fractional. It is
-    /// recommended to calculate sizes as follows:
-    /// ```
-    /// use kas_core::cast::*;
-    /// # let scale_factor = 1.5f32;
-    /// let size: i32 = (100.0 * scale_factor).cast_ceil();
-    /// ```
-    ///
-    /// This value may change during a program's execution (e.g. when a window
-    /// is moved to a different monitor); in this case all widgets will be
-    /// resized via [`crate::Layout::size_rules`].
-    fn scale_factor(&self) -> f32;
-
-    /// Convert a size in virtual pixels to physical pixels
-    fn pixels_from_virtual(&self, px: f32) -> f32 {
-        px * self.scale_factor()
-    }
-
-    /// Convert a size in font Points to physical pixels
-    fn pixels_from_points(&self, pt: f32) -> f32;
-
-    /// Convert a size in font Em to physical pixels
-    ///
-    /// (This depends on the font size.)
-    fn pixels_from_em(&self, em: f32) -> f32;
-
-    /// Size of a frame around child widget(s)
-    ///
-    /// This already includes the margins specified by [`Self::frame_margins`].
-    fn frame(&self, vert: bool) -> FrameRules;
-
-    /// Frame/margin around a menu entry
-    fn menu_frame(&self, vert: bool) -> FrameRules;
-
-    /// Size of a separator frame between items
-    fn separator(&self) -> Size;
-
-    /// Size of a navigation highlight margin around a child widget
-    fn nav_frame(&self, vert: bool) -> FrameRules;
-
-    /// The margin around content within a widget
-    ///
-    /// Though inner margins are *usually* empty, they are sometimes drawn to,
-    /// for example focus indicators.
-    fn inner_margin(&self) -> Size;
-
-    /// The margin between UI elements, where desired
-    ///
-    /// Widgets must not draw in outer margins.
-    fn outer_margins(&self) -> Margins;
-
-    /// The margin around frames and separators
-    fn frame_margins(&self) -> Margins;
-
-    /// The margin around text elements
-    ///
-    /// Similar to [`Self::outer_margins`], but intended for things like text
-    /// labels which do not have a visible hard edge.
-    fn text_margins(&self) -> Margins;
-
-    /// The height of a line of text
-    fn line_height(&self, class: TextClass) -> i32;
-
-    /// Update a [`Text`] and get a size bound
-    ///
-    /// First, this method updates the text's [`Environment`]: `bounds`, `dpp`
-    /// and `pt_size` are set. Second, the text is prepared (which is necessary
-    /// to calculate size requirements). Finally, this converts the requirements
-    /// to a [`SizeRules`] value and returns it.
-    ///
-    /// Usually this method is used in [`Layout::size_rules`], then
-    /// [`TextApiExt::update_env`] is used in [`Layout::set_rect`].
-    ///
-    /// [`Environment`]: crate::text::Environment
-    /// [`Layout::set_rect`]: crate::Layout::set_rect
-    /// [`Layout::size_rules`]: crate::Layout::size_rules
-    fn text_bound(&self, text: &mut dyn TextApi, class: TextClass, axis: AxisInfo) -> SizeRules;
-
-    /// Width of an edit marker
-    fn edit_marker_width(&self) -> f32;
-
-    /// Size of the sides of a button.
-    fn button_surround(&self, vert: bool) -> FrameRules;
-
-    /// Size of the frame around an edit box, including margin
-    ///
-    /// Note: though text should not be drawn in the margin, the edit cursor
-    /// may be. The margin included here should be large enough!
-    fn edit_surround(&self, vert: bool) -> FrameRules;
-
-    /// Size of the element drawn by [`DrawHandle::checkbox`].
-    fn checkbox(&self) -> Size;
-
-    /// Size of the element drawn by [`DrawHandle::radiobox`].
-    fn radiobox(&self) -> Size;
-
-    /// Dimensions for a scrollbar
-    ///
-    /// Returns:
-    ///
-    /// -   `size`: minimum size of handle in horizontal orientation;
-    ///     `size.1` is also the width of the scrollbar
-    /// -   `min_len`: minimum length for the whole bar
-    ///
-    /// Required bound: `min_len >= size.0`.
-    fn scrollbar(&self) -> (Size, i32);
-
-    /// Dimensions for a slider
-    ///
-    /// Returns:
-    ///
-    /// -   `size`: minimum size of handle in horizontal orientation;
-    ///     `size.1` is also the width of the slider
-    /// -   `min_len`: minimum length for the whole bar
-    ///
-    /// Required bound: `min_len >= size.0`.
-    fn slider(&self) -> (Size, i32);
-
-    /// Dimensions for a progress bar
-    ///
-    /// Returns the minimum size for a horizontal progress bar. It is assumed
-    /// that the width is adjustable while the height is (preferably) not.
-    /// For a vertical bar, the values are swapped.
-    fn progress_bar(&self) -> Size;
-}
 
 /// A handle to the active theme, used for drawing
 ///
@@ -493,77 +242,6 @@ pub trait DrawHandleExt: DrawHandle {
 impl<D: DrawHandle + ?Sized> DrawHandleExt for D {}
 
 macro_rules! impl_ {
-    (($($args:tt)*) SizeHandle for $ty:ty) => {
-        impl<$($args)*> SizeHandle for $ty {
-            fn scale_factor(&self) -> f32 {
-                self.deref().scale_factor()
-            }
-            fn pixels_from_points(&self, pt: f32) -> f32 {
-                self.deref().pixels_from_points(pt)
-            }
-            fn pixels_from_em(&self, em: f32) -> f32 {
-                self.deref().pixels_from_em(em)
-            }
-
-            fn frame(&self, vert: bool) -> FrameRules {
-                self.deref().frame(vert)
-            }
-            fn menu_frame(&self, vert: bool) -> FrameRules {
-                self.deref().menu_frame(vert)
-            }
-            fn separator(&self) -> Size {
-                self.deref().separator()
-            }
-            fn nav_frame(&self, vert: bool) -> FrameRules {
-                self.deref().nav_frame(vert)
-            }
-            fn inner_margin(&self) -> Size {
-                self.deref().inner_margin()
-            }
-            fn outer_margins(&self) -> Margins {
-                self.deref().outer_margins()
-            }
-            fn frame_margins(&self) -> Margins {
-                self.deref().frame_margins()
-            }
-            fn text_margins(&self) -> Margins {
-                self.deref().text_margins()
-            }
-
-            fn line_height(&self, class: TextClass) -> i32 {
-                self.deref().line_height(class)
-            }
-            fn text_bound(&self, text: &mut dyn TextApi, class: TextClass, axis: AxisInfo) -> SizeRules {
-                self.deref().text_bound(text, class, axis)
-            }
-            fn edit_marker_width(&self) -> f32 {
-                self.deref().edit_marker_width()
-            }
-
-            fn button_surround(&self, vert: bool) -> FrameRules {
-                self.deref().button_surround(vert)
-            }
-            fn edit_surround(&self, vert: bool) -> FrameRules {
-                self.deref().edit_surround(vert)
-            }
-
-            fn checkbox(&self) -> Size {
-                self.deref().checkbox()
-            }
-            fn radiobox(&self) -> Size {
-                self.deref().radiobox()
-            }
-            fn scrollbar(&self) -> (Size, i32) {
-                self.deref().scrollbar()
-            }
-            fn slider(&self) -> (Size, i32) {
-                self.deref().slider()
-            }
-            fn progress_bar(&self) -> Size {
-                self.deref().progress_bar()
-            }
-        }
-    };
     (($($args:tt)*) DrawHandle for $ty:ty) => {
         impl<$($args)*> DrawHandle for $ty {
             fn size_handle(&self) -> &dyn SizeHandle {
@@ -662,8 +340,6 @@ macro_rules! impl_ {
         }
     };
 }
-
-impl_! { (S: SizeHandle + ?Sized, R: Deref<Target = S>) SizeHandle for R }
 
 impl_! { (H: DrawHandle) DrawHandle for Box<H> }
 #[cfg(feature = "stack_dst")]
