+++ crates/kas-theme/src/theme_dst.rs
@@ -10,9 +10,8 @@ use std::borrow::Cow;
 use std::ops::{Deref, DerefMut};
 
 use super::{StackDst, Theme, Window};
-use kas::draw::{
-    color, DrawHandle, DrawIface, DrawSharedImpl, SharedState, SizeHandle, ThemeControl,
-};
+use kas::draw::{color, DrawIface, DrawSharedImpl, SharedState};
+use kas::theme::{DrawHandle, SizeHandle, ThemeControl};
 use kas::TkAction;
 
 /// An optionally-owning (boxed) reference
